Github login; implemented API authentication

+++ db/migrate/20160813210100_add_facebook_id_to_users.rb
@@ -1,5 +1,5 @@
 class AddFacebookIdToUsers < ActiveRecord::Migration[5.0]
   def change
-    add_column :users, :facebook_id, :integer
+    add_column :users, :facebook_id, :bigint
   end
 end
